+++ src/ast/scopes/GlobalScope.ts
@@ -1,8 +1,14 @@
 import GlobalVariable from '../variables/GlobalVariable';
+import UndefinedVariable from '../variables/UndefinedVariable';
 import Scope from './Scope';
 
 export default class GlobalScope extends Scope {
 	parent: null;
+
+	constructor() {
+		super();
+		this.variables.undefined = new UndefinedVariable();
+	}
 
 	findVariable(name: string) {
 		if (!this.variables[name]) {