+++ test/function/samples/allow-undefined-as-parameter/_config.js
@@ -1,0 +1,3 @@
+module.exports = {
+	description: 'do not make assumptions about its value when using undefined as a parameter name'
+};