+++ test/function/samples/does-not-deconflict-undefined/_config.js
@@ -1,0 +1,3 @@
+module.exports = {
+	description: 'does not rename undefined'
+};